(PC-11746) refactor(BeneficiaryRequestSent): move screen to IdentityCheck
