--- src/features/auth/signup/BeneficiaryRequestSent/index.tsx
@@ -1,1 +1,0 @@
-export * from './BeneficiaryRequestSent'
