+++ src/features/identityCheck/pages/confirmation/BeneficiaryRequestSent.tsx
@@ -28,22 +28,23 @@ export function BeneficiaryRequestSent() {
     }
   }
 
-  let body = t`Tu recevras un e-mail lorsque ta demande sera validée.`
+  const body = settings?.enableIdCheckRetention
+    ? t`Tu recevras une réponse par e-mail sous 5 jours ouvrés.`
+    : t`Tu recevras un e-mail lorsque ta demande sera validée.`
+
+  let inTheMeantime = ''
   if (shouldNavigateToCulturalSurvey) {
-    body = t`Tu recevras un e-mail lorsque ta demande sera validée. En attendant, aide-nous à en savoir plus sur tes pratiques culturelles !`
+    inTheMeantime = t`En attendant, aide-nous à en savoir plus sur tes pratiques culturelles !`
+  } else if (settings?.enableIdCheckRetention) {
+    inTheMeantime = t`En attendant, tu peux découvrir l'application !`
   }
 
-  if (settings?.enableIdCheckRetention) {
-    body = t`Tu recevras une réponse par e-mail sous 5 jours ouvrés. En attendant, tu peux découvrir l'application !`
-    if (shouldNavigateToCulturalSurvey) {
-      body = t`Tu recevras une réponse par e-mail sous 5 jours ouvrés. En attendant, aide-nous à en savoir plus sur tes pratiques culturelles !`
-    }
-  }
+  const message = inTheMeantime.length ? `${body} ${inTheMeantime}` : body
 
   return (
     <GenericInfoPage title={t`Demande envoyée !`} icon={RequestSent} iconSize={getSpacing(42)}>
       <StyledBody>{t`Nous étudions ton dossier...`}</StyledBody>
-      <StyledBody>{body}</StyledBody>
+      <StyledBody>{message}</StyledBody>
       <Spacer.Column numberOfSpaces={15} />
       <ButtonPrimaryWhite title={t`On y va !`} onPress={onPress} />
     </GenericInfoPage>
